decode/encode functions moved to StartDateTimePicker
